Add solution for exercise 4
Mostly shows the need to improve on non-repetition. Title does not to
be a flex box - interesting trick problem. Therefore `flex` and the
alignment only need to be done for the `.main` container. Don't need
`flex-direction: column` because the text wraps due to the set width.
Note to self to list similarities between all items first.